+++ public/components/react-island/react-island--collect-consents.js
@@ -3,7 +3,7 @@ const selector: string = '.react-island--collect-consents';
 
 const init = ($component: HTMLElement): void => {
   Promise.all([
-    import('elements/CollectConsents'),
+    import('react-elements/CollectConsents'),
     import('js/hydrate-react-island')
   ]).then(([{ CollectConsents }, { hydrate }]) => {
     hydrate($component, CollectConsents);
